feat: un seul btn supprimer entrepôt, nettoyage puis demande de suppr
